fix: image compression logic
